tests/wpt/tests/css/css-conditional/container-queries/scroll-state-snapped-container-type-change.html

<!DOCTYPE html>
<title>@container: scroll-state(snapped) property changes</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<style>
  :root {
    scroll-snap-type: block mandatory;
  }
  body {
    margin: 0;
  }
  #filler-before {
    height: 200px;
  }
  #filler-after {
    height: 10000px;
  }
  #snapped {
    container-type: scroll-state;
    scroll-snap-align: start;
    width: 100px;
    height: 100px;
    background: teal;
  }
  span {
    --snapped: no;
    @container scroll-state(snapped) {
      --snapped: yes;
    }
  }
</style>
<div id="filler-before"></div>
<div id="snapped">
  <span id="target">My container is snapped</span>
</div>
<div id="filler-after"></div>
<script>
  setup(() => assert_implements_container_queries());

  promise_test(async t => {
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped"), "yes");
  }, "Check scroll-state(snapped) initially matching");

  promise_test(async t => {
    t.add_cleanup(async () => snapped.style.scrollSnapAlign = "");
    snapped.style.scrollSnapAlign = "initial";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped"), "no",
                  "scroll-snap-align removed");
    snapped.style.scrollSnapAlign = "";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped"), "yes",
                  "scroll-snap-align re-added");
  }, "Check scroll-state(snapped) not matching when scroll-snap-align is removed");

  promise_test(async t => {
    t.add_cleanup(async () => document.documentElement.style.scrollSnapType = "");
    document.documentElement.style.scrollSnapType = "initial";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped"), "no",
                  "scroll-snap-type removed");
    document.documentElement.style.scrollSnapType = "";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped"), "yes",
                  "scroll-snap-type re-added");
  }, "Check scroll-state(snapped) not matching when scroll-snap-type is removed");

  promise_test(async t => {
    t.add_cleanup(async () => snapped.style.containerType = "");
    snapped.style.containerType = "initial";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped"), "no",
                  "container-type removed");
    snapped.style.containerType = "";
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target).getPropertyValue("--snapped"), "yes",
                  "container-type re-added");
  }, "Check scroll-state(snapped) not matching when container-type is removed");

</script>
